Chem.is.try V3 / Gaia Updates
Updated Gaia to have a Object prototype that iterates through an objects
properties and list the corresponding values.

For Chem.is.try V3 I converted the data structure to JSON and something
that made more sense.

There is now a data explorer application to easily view and update the
data.

There is a converter to go from the old format to the new.

In the old format an entry for pretty formulas used to be in the data
structure. Now there is a prettyFormula function that does the work
after the fact.